dom/base/test/test_meta_viewport_negative_width_and_no_height.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>negative width in meta viewport</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <meta name="viewport" content="width=-400">
  <script src="viewport_helpers.js"></script>
</head>
<body>
  <p>width=-400</p>
  <script type="application/javascript">
    "use strict";

    add_task(async function negative_width() {
      await SpecialPowers.pushPrefEnv(scaleRatio(1.0));

      let info = getViewportInfo(800, 480);
      is(info.width, 800, "width should be 800");
      is(info.height, 480, "height should be 480");
    });
  </script>
</body>
</html>
